requirements.txt
groq
python-dotenv==1.0.0
streamlit>=1.36
reportlab>=4.2
